use bigint for filesize

+++ install/create_tables-postgres.sql
@@ -223,7 +223,7 @@ CREATE TABLE "tblDocumentContent" (
   "orgFileName" varchar(150) NOT NULL default '',
   "fileType" varchar(10) NOT NULL default '',
   "mimeType" varchar(100) NOT NULL default '',
-  "fileSize" INTEGER,
+  "fileSize" BIGINT,
   "checksum" char(32),
   UNIQUE ("document","version")
 ) ;
